Update Express middleware configuration
