+++ server/router.js
@@ -1,5 +1,6 @@
 const express = require('express');
 const router = express.Router();
+const { decryptURI } = require('./controllers/CryptoJSController');
 const { getSQLTables, getSQLDBname, prepForGQL } = require('./controllers/SQLController');
 const { connectToMongo, getMongoDocuments, MongoPrepForGQL } = require('./controllers/MongoDBController');
 const { convertToGQLServerCode } = require('./controllers/GQLServerController');
@@ -12,6 +13,7 @@ router.get('/', (req, res) => {
 /* ROUTE TO GET DEMO POSTGRESQL DB AND CONVERT TO GRAPHQL SERVER, CLIENT QUERIES AND MUTATIONS CODE */
 router.get(
   '/convert-demo-db',
+  decryptURI,
   getSQLDBname,
   getSQLTables,
   prepForGQL,
@@ -19,11 +21,6 @@ router.get(
   convertToGQLClientQueriesCode,
   convertToGQLClientMutationsCode,
   (req, res) => {
-    // console.log('this is res.locals', res.locals);
-    // console.log('this is res.locals.GQLServerCode', res.locals.GQLServerCode);
-    // console.log('this is res.locals.GQLClientQueriesCode', res.locals.GQLClientQueriesCode);
-    // console.log('this is res.locals.GQLClientMutationsCode', res.locals.GQLClientMutationsCode);
-
     const GQLCode = {
       DBName: res.locals.DBname,
       SQLSchema: res.locals.SQLSchema,
@@ -38,6 +35,7 @@ router.get(
 /* ROUTE TO GET USER POSTGRESQL DB AND CONVERT TO GRAPHQL SERVER, CLIENT QUERIES AND MUTATIONS CODE */
 router.post(
   '/convert-sql-db',
+  decryptURI,
   getSQLDBname,
   getSQLTables,
   prepForGQL,
@@ -45,11 +43,6 @@ router.post(
   convertToGQLClientQueriesCode,
   convertToGQLClientMutationsCode,
   (req, res) => {
-    // console.log('this is res.locals', res.locals);
-    // console.log('this is res.locals.GQLServerCode', res.locals.GQLServerCode);
-    // console.log('this is res.locals.GQLClientQueriesCode', res.locals.GQLClientQueriesCode);
-    // console.log('this is res.locals.GQLClientMutationsCode', res.locals.GQLClientMutationsCode);
-
     const GQLCode = {
       DBName: res.locals.DBname,
       SQLSchema: res.locals.SQLSchema,
@@ -64,6 +57,7 @@ router.post(
 /* ROUTE TO GET USER MONGO DB AND CONVERT TO GRAPHQL SERVER, CLIENT QUERIES AND MUTATIONS CODE */
 router.post(
   '/convert-mongo-db',
+  decryptURI,
   connectToMongo,
   getMongoDocuments,
   MongoPrepForGQL,
@@ -71,11 +65,6 @@ router.post(
   convertToGQLClientQueriesCode,
   convertToGQLClientMutationsCode,
   (req, res) => {
-    // console.log('this is res.locals', res.locals);
-    // console.log('this is res.locals.GQLServerCode', res.locals.GQLServerCode);
-    // console.log('this is res.locals.GQLClientQueriesCode', res.locals.GQLClientQueriesCode);
-    // console.log('this is res.locals.GQLClientMutationsCode', res.locals.GQLClientMutationsCode);
-
     const GQLCode = {
       DBName: res.locals.DBname,
       MongoSchema: res.locals.MongoSchema,
@@ -83,7 +72,6 @@ router.post(
       GQLClientQueriesCode: res.locals.GQLClientQueriesCode,
       GQLClientMutationsCode: res.locals.GQLClientMutationsCode,
     };
-    // res.status(200).json(GQLCode);
     res.status(200).json(GQLCode);
   }
 );
